+++ .idea/misc.xml
@@ -26,6 +26,8 @@
         <entry key="..\:/Users/82103/Documents/GitHub/app/src/main/res/layout/normal_fragment.xml" value="0.24974358974358973" />
         <entry key="..\:/Users/82103/Documents/GitHub/app/src/main/res/menu/bottom_nav_menu.xml" value="0.25364583333333335" />
         <entry key="..\:/Users/82103/Documents/GitHub/app/src/main/res/menu/nav_menu.xml" value="0.25364583333333335" />
+        <entry key="..\:/Users/GyeongSeok/Documents/GitHub/CAPs-1/app/src/main/res/layout/activity_login.xml" value="0.358695652173913" />
+        <entry key="..\:/Users/GyeongSeok/Documents/GitHub/CAPs-1/app/src/main/res/layout/activity_main.xml" value="0.358695652173913" />
         <entry key="..\:/Users/rudtj/Documents/GitHub/CAPs/app/src/main/res/drawable-v24/ic_launcher_foreground.xml" value="0.37604166666666666" />
         <entry key="..\:/Users/rudtj/Documents/GitHub/CAPs/app/src/main/res/drawable/ic_baseline_account_circle_24.xml" value="0.37604166666666666" />
         <entry key="..\:/Users/rudtj/Documents/GitHub/CAPs/app/src/main/res/drawable/ic_baseline_arrow_back_60.xml" value="0.18958333333333333" />
