+++ app/src/main/java/com/example/cpas/home/NormalFragment.kt
@@ -1,18 +1,17 @@
-package com.example.cpas
+package com.example.cpas.home
 
-import android.app.ActionBar
 import android.os.Bundle
-import android.util.DisplayMetrics
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
-import android.widget.TextView
 import android.widget.Toast
 import androidx.fragment.app.Fragment
 import androidx.recyclerview.widget.LinearLayoutManager
 import androidx.recyclerview.widget.RecyclerView
+import com.example.cpas.posting.Posting
+import com.example.cpas.posting.PostingAdapter
+import com.example.cpas.R
 import com.google.firebase.database.*
-import java.util.*
 import kotlin.Comparator
 import kotlin.collections.ArrayList
