+++ app/src/main/java/com/example/cpas/home/PagerAdapter.kt
@@ -1,4 +1,4 @@
-package com.example.cpas
+package com.example.cpas.home
 
 import androidx.fragment.app.Fragment
 import androidx.fragment.app.FragmentActivity
