+++ app/src/main/java/com/example/cpas/planner/PlanData.kt
@@ -1,3 +1,3 @@
-package com.example.cpas
+package com.example.cpas.planner
 
 data class planData(val title : String, val time : String)
