+++ app/src/main/java/com/example/cpas/posting/Comment.kt
@@ -1,3 +1,3 @@
-package com.example.cpas
+package com.example.cpas.posting
 
 data class Comment(val who : String, val id : String, val content : String, val time : String, val epoch : String, val commentID : String)
