+++ app/src/main/java/com/example/cpas/posting/Posting.kt
@@ -1,3 +1,3 @@
-package com.example.cpas
+package com.example.cpas.posting
 
 data class Posting(val id:String?, val type : String?, val title : String, val content : String, val time : String, val commentNum : String, val who:String?, val commentImage : Int, val epoch : String?, val postingID : String?)
