+++ app/src/main/java/com/example/cpas/posting/Postings.kt
@@ -1,3 +1,3 @@
-package com.example.cpas
+package com.example.cpas.posting
 
 data class Postings(val id:String?, val type : String?, val title : String, val content : String, val time : String, val commentNum : Int, val who:String?)
