+++ app/src/main/res/layout/activity_assign.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".AssignActivity">
+    tools:context=".assign.AssignActivity">
 
     <TextView
         android:id="@+id/AssTitle"
