+++ app/src/main/res/layout/activity_in_posting.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".InPosting">
+    tools:context=".posting.InPosting">
 
     <ScrollView
         android:layout_width="match_parent"
