+++ app/src/main/res/layout/activity_set_plan.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".SetPlanActivity">
+    tools:context=".planner.SetPlanActivity">
 
     <ImageButton
         android:id="@+id/cancel_button"
